Fix pattern.cc using default int
This is causing us to fail to build on windows_x64
because int is getting defaulted to 64 bit and
when we call the ICU functions, we are down casting
the type and risking losing info.

Bug: 4894239

+++ base/strings/pattern.cc
@@ -123,7 +123,7 @@ constexpr bool MatchPatternT(const CHAR* eval,
 struct NextCharUTF8 {
   base_icu::UChar32 operator()(const char** p, const char* end) {
     base_icu::UChar32 c;
-    int offset = 0;
+    int32_t offset = 0;
     CBU8_NEXT(reinterpret_cast<const uint8_t*>(*p), offset, end - *p, c);
     *p += offset;
     return c;
@@ -133,7 +133,7 @@ struct NextCharUTF8 {
 struct NextCharUTF16 {
   base_icu::UChar32 operator()(const char16_t** p, const char16_t* end) {
     base_icu::UChar32 c;
-    int offset = 0;
+    size_t offset = 0;
     CBU16_NEXT(*p, offset, end - *p, c);
     *p += offset;
     return c;
